bugfix: Changing HDD controller for Hardfile lost the path to the hardfile
The GUI had a bug where the Hardfile select dialog would lose the path, if you changeed the HDD controller type

+++ src/osdep/gui/PanelHD.cpp
@@ -81,7 +81,7 @@ public:
 
 	int getNumberOfElements() override
 	{
-		return lstMRUCDList.size();
+		return static_cast<int>(lstMRUCDList.size());
 	}
 
 	int add_element(const char* elem) override
